+++ carcassonedraw.cbp
@@ -39,9 +39,18 @@
 		<Compiler>
 			<Add option="-Wall" />
 		</Compiler>
+		<Unit filename="essential.h" />
+		<Unit filename="graphics.c">
+			<Option compilerVar="CC" />
+		</Unit>
+		<Unit filename="graphics.h" />
 		<Unit filename="source.c">
 			<Option compilerVar="CC" />
 		</Unit>
+		<Unit filename="textinterface.c">
+			<Option compilerVar="CC" />
+		</Unit>
+		<Unit filename="textinterface.h" />
 		<Extensions>
 			<code_completion />
 			<envvars />
