+++ textinterface.c
@@ -1,0 +1,14 @@
+#include "textinterface.h"
+
+void TEXT_interface_init()
+{
+
+}
+void TEXT_user_loop()
+{
+
+}
+void TEXT_end_application()
+{
+
+}
