+++ textinterface.h
@@ -1,0 +1,11 @@
+#ifndef TEXTINTERFACE_H_INCLUDED
+#define TEXTINTERFACE_H_INCLUDED
+#include "essential.h"
+
+void TEXT_interface_init();
+void TEXT_user_loop();
+void TEXT_end_application();
+
+//tu funkcje, ktorych uzywac bedzie tylko interfejs tekstowy
+
+#endif // TEXTINTERFACE_H_INCLUDED
